+++ iOS-HW17-Olga Mikhailova2/ViewController.swift
@@ -1,0 +1,19 @@
+//
+//  ViewController.swift
+//  iOS-HW17-Olga Mikhailova2
+//
+//  Created by FoxxFire on 13.06.2025.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view.
+    }
+
+
+}
+
